+++ src/methods/$last-or/$last-or.d.ts
@@ -1,7 +1,0 @@
-import { $Promise } from '../../../generate/async.macro';
-
-import { $SourceIterable } from '../../types/$iterable';
-
-declare function $lastOr<E, T>(whenEmpty: E, iterable: $SourceIterable<T>): $Promise<T | E>;
-
-export default $lastOr;